+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
 
+### Version 1.0.1
+- **Enhancement**: Added category `description` and `gc_metadata` attributes to the sample data setup.
+- **Enhancement**: Added `media gallery assets` to the sample data setup.
+
 ### Version 1.0.0
 - **Feature**: New module used for sample data.